third_party/blink/web_tests/fast/css/target-fragment-match-expected.txt

Verify that css :target selector is correctly updated during hash and history navigations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById('target-01')).backgroundColor is "rgb(102, 204, 255)"
PASS getComputedStyle(document.getElementById('target-01')).backgroundColor is "rgb(255, 255, 255)"
PASS successfullyParsed is true

TEST COMPLETE


I should be highlighted first because of the anchor, and de-highlighted when there is no fragment.
